app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
aiosqlite==0.19.0
pydantic==2.5.0
python-docx==1.1.0
PyPDF2==3.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
huggingface-hub==0.25.2
